refactor(discovery): split into common / announcing / browsing modules (#90)
* refactor(discovery): split into common / announcing / browsing modules

Pure refactor — no behavior change. Splits discovery.rs into a sibling
module layout so the daemon setup, announcing path, and browsing path
each live in their own file, with shared helpers and the peer registry
in common.rs.

```
discovery.rs       orchestrator: MdnsHandle, publish() entry point,
                   mdns_sd_start wires the three submodules together
discovery/common.rs    constants, name builders, IP helpers,
                       mdns-sd daemon setup + interface filter,
                       PeerRegistry
discovery/announcing.rs  mdns_sd_register
discovery/browsing.rs    mdns_sd_start + ServiceEvent handling
```


* docs: add ADR-014 — split discovery.rs into common / announcing / browsing submodules

---------
