+++ documents/adr/ADR-014.md
@@ -1,0 +1,66 @@
+# ADR-014: Split `discovery.rs` into `common` / `announcing` / `browsing` submodules
+
+**Status:** Accepted
+**Date:** 28-05-2026
+**Author:** Yordan Mitev
+
+## Context
+
+ADR-013 settled the DNS-SD backend on `mdns-sd` for every platform. After that landed, the resulting `discovery.rs` was a single ~500-line file that mixed five unrelated concerns:
+
+* On-the-wire identity (service type, TXT keys, instance / SRV name builders).
+* LAN-IP detection and the loopback / link-local routability check.
+* `mdns-sd` daemon construction and the virtual-interface filter.
+* The publishing path (`mdns_sd_register` — `ServiceInfo::new`, props, register call).
+* The browsing path (`mdns_sd_start` — `ServiceDaemon::browse`, event loop, `ServiceEvent` handlers, endpoint translation).
+
+Two follow-up PRs were already in flight against this file — the session filter (PR #91) needs touch sites in both the publish and browse paths, and the agent-inventory work (PR #92) interacts with the peer registry. Both reviewers complained that diffs against a single 500-line module are hard to read because every change appears far from the others it relates to.
+
+The publish and browse halves also have different test surfaces: the publish path is exercised by interaction with the mdns-sd daemon (real socket); the browse path is exercised by handcrafted `ServiceEvent` values. Keeping them in one file makes it impossible to attach test modules to either half without dragging in the other.
+
+## Decision
+
+Split `discovery.rs` into a sibling module layout:
+
+```
+src/zenoh_router/src/discovery.rs           orchestrator
+src/zenoh_router/src/discovery/common.rs    shared constants + helpers
+src/zenoh_router/src/discovery/announcing.rs publish path
+src/zenoh_router/src/discovery/browsing.rs   browse path
+```
+
+* `discovery.rs` keeps the public surface (`MdnsHandle`, `MdnsHandle::publish`) and the `mdns_sd_start` glue that wires the three submodules together. Drop on the handle still unregisters the service and shuts the daemon down.
+* `common.rs` owns the on-the-wire identity (`SERVICE_TYPE`, `TXT_KEY_CN`, `TXT_KEY_IP`), the name builders (`instance_name`, `srv_host_name`, `zenoh_tls_endpoint`), the LAN-IP detection trick (`detect_lan_ipv4`, `is_unroutable`), the daemon construction + interface filter, and the `PeerRegistry` watch-channel wrapper used by the browser.
+* `announcing.rs` owns `mdns_sd_register` — building `ServiceInfo`, pinning the A record to the LAN IPv4, calling `daemon.register`.
+* `browsing.rs` owns `mdns_sd_start` — calling `daemon.browse`, draining `ServiceEvent`s, resolving each peer into a list of `tls/<ip>:<port>` endpoints, pushing into `PeerRegistry`.
+
+Visibility uses `pub(super)` throughout: each submodule exposes only what the orchestrator and its siblings need. Nothing leaks above the `discovery` module boundary that isn't already in the public `MdnsHandle` API.
+
+This is a **pure refactor** — no behavior change. The on-the-wire DNS-SD output (PTR / SRV / TXT shape, instance / SRV target naming, TXT keys) is byte-identical before and after.
+
+## Alternatives
+
+**Keep one file.** Rejected. Already painful at 500 lines; PR #91 adds ~80 lines spread across publish + browse + state, PR #92 adds ~150 more in similar pattern. Reviewing those PRs against the single file is exactly the failure case this split solves.
+
+**Split by backend instead of by direction (e.g. `mdns_sd.rs`, with future `bonjour.rs` etc.).** Rejected. ADR-013 commits us to a single backend; planning the file layout around a future second backend that we have explicitly decided not to build adds permanent structural weight for a hypothetical that may never materialise.
+
+**Extract the peer registry to its own crate.** Rejected. The registry is a thin `HashMap<String, Vec<String>>` + `watch::Sender` wrapper, tightly coupled to the browse loop's event semantics. A separate crate would mean a published versioning story for ~50 lines of glue. Co-located in `common.rs` is appropriate.
+
+**Use one `mod.rs`-style directory with `discovery/mod.rs`.** Rejected. Rust 2018+ idiom and this codebase's convention both prefer the `name.rs` + `name/` sibling layout. Consistency matters more than the small layout difference.
+
+## Consequences
+
+**Positive**
+* Each file has one clearly-stated purpose, named in its module-doc comment. New contributors can find the publish vs. browse logic without grepping.
+* Subsequent PRs (#91 session filter, #92 agent inventory) touch one or two files each instead of one large one. Reviewer diff comprehension improves linearly with the split.
+* Test modules can be attached per-direction in a follow-up — the publish path's tests no longer have to drag the browse path through their compile graph.
+* `discovery.rs`'s top-level module-doc now reads as a table of contents pointing at the three submodules, which is the documentation the previous file was missing.
+
+**Negative**
+* `git blame` discontinuity at the split commit. Mitigated because the commit is a pure refactor with a clear message; `git log --follow` continues to work for the migrated symbols.
+* One extra layer of indirection when reading the publish or browse path top-to-bottom: starting from `MdnsHandle::publish` now requires two jumps (orchestrator → submodule) instead of one. The trade-off is acceptable given the readability gain at scale.
+* Future cross-cutting changes (e.g. swapping the wire format) now touch three files instead of one. This is the cost of separation-of-concerns and is the same trade made everywhere else in the workspace.
+
+---
+
+*Artificial intelligence was used to support the drafting of parts of this document. The generated text was subsequently reviewed, checked, and edited where necessary by the author, Yordan Mitev.*
